+++ cmd/exec.go
@@ -39,8 +39,27 @@ before the command gets executed in each directory.`,
 	cmd.Flags().StringSliceVarP(&tags, "tags", "t", []string{}, "target projects by their tag")
 	cmd.Flags().StringSliceVarP(&projects, "projects", "p", []string{}, "target projects by their name")
 
-	cmd.MarkFlagCustom("projects", "__mani_parse_projects")
-	cmd.MarkFlagCustom("tags", "__mani_parse_tags")
+	cmd.RegisterFlagCompletionFunc("projects", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		_, config, err := core.ReadConfig(*configFile)
+
+		if err != nil {
+			return []string{}, cobra.ShellCompDirectiveDefault
+		}
+
+		projects := core.GetProjectNames(config.Projects)
+		return projects, cobra.ShellCompDirectiveDefault
+	})
+
+	cmd.RegisterFlagCompletionFunc("tags", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		_, config, err := core.ReadConfig(*configFile)
+
+		if err != nil {
+			return []string{}, cobra.ShellCompDirectiveDefault
+		}
+
+		tags := core.GetTags(config.Projects)
+		return tags, cobra.ShellCompDirectiveDefault
+	})
 
 	return &cmd
 }