+++ test/integration/completion_test.go
@@ -14,27 +14,27 @@ var completionTests = []TemplateTest{
 	},
 
 	{
-		TestName: "Print zsh completion",
-		InputFiles: []string { "" },
-		TestCmd: "$MANI completion zsh",
-		Golden: "completion/zsh",
-		WantErr: false,
+		TestName:   "Print zsh completion",
+		InputFiles: []string{},
+		TestCmd:    "$MANI completion zsh",
+		Golden:     "completion/zsh",
+		WantErr:    false,
 	},
 
 	{
-		TestName: "Print fish completion",
-		InputFiles: []string { "" },
-		TestCmd: "$MANI completion fish",
-		Golden: "completion/fish",
-		WantErr: false,
+		TestName:   "Print fish completion",
+		InputFiles: []string{},
+		TestCmd:    "$MANI completion fish",
+		Golden:     "completion/fish",
+		WantErr:    false,
 	},
 
 	{
-		TestName: "Print powershell completion",
-		InputFiles: []string { "" },
-		TestCmd: "$MANI completion psh",
-		Golden: "completion/psh",
-		WantErr: false,
+		TestName:   "Print powershell completion",
+		InputFiles: []string{},
+		TestCmd:    "$MANI completion powershell",
+		Golden:     "completion/powershell",
+		WantErr:    false,
 	},
 }